[Update]update the project name from  to
