+++ rocketmq-replicator/src/main/java/org/apache/rocketmq/replicator/config/DataType.java
@@ -14,10 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.rocketmq.connector.strategy;
+package org.apache.rocketmq.replicator.config;
 
-public enum DivideStrategyEnum {
+public enum DataType {
 
-    BY_TOPIC,
-    BY_QUEUE
+    COMMON_MESSAGE,
+    TOPIC_CONFIG,
+    BROKER_CONFIG,
+    SUB_CONFIG
 }
